refactor mcp rule tool schemas
